% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_table.R
\name{extract_table}
\alias{extract_table}
\title{Extract a table from a character vector}
\usage{
extract_table(
  tbl_vctr = character(),
  merge_empty = TRUE,
  orientation = c("horizontal", "vertical")
)
}
\arguments{
\item{tbl_vctr}{A character vector with one line of text per value.}

\item{merge_empty}{Whether to merge empty cells together (by "row" then by
"column").}

\item{orientation}{Whether the data is laid out horizontally like a tidy
table (variables = column names), or vertically (more like a list).}
}
\value{
A data.frame with the extracted table.
}
\description{
This is likely not the final form and/or name of this function. Right now it
requires a character vector that has already been split by newlines, and some
guidance on what to do. It does well once you tell it where to look, though.
}
