% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableguess-package.R
\docType{package}
\name{tableguess-package}
\alias{tableguess}
\alias{tableguess-package}
\title{tableguess: Find Tables in Plain Text}
\description{
The goal of tableguess is to...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/tableguess}
  \item \url{https://jonthegeek.github.io/tableguess/}
  \item Report bugs at \url{https://github.com/jonthegeek/tableguess/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
